+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.0.0
+
+- Remove repository parameter
+- Add repositories parameter, allow list of comma separated
+
 ## 1.0.0
 
 - Initial version